Refactor internal package of ModifierExtensions file
