+++ stream-video-android-compose/src/main/kotlin/io/getstream/video/android/compose/ui/components/call/ringing/outgoingcall/OutgoingCallControls.kt
@@ -33,7 +33,7 @@ import io.getstream.video.android.compose.theme.VideoTheme
 import io.getstream.video.android.compose.ui.components.call.controls.actions.CancelCallAction
 import io.getstream.video.android.compose.ui.components.call.controls.actions.ToggleCameraAction
 import io.getstream.video.android.compose.ui.components.call.controls.actions.ToggleMicrophoneAction
-import io.getstream.video.android.compose.ui.components.extensions.toggleAlpha
+import io.getstream.video.android.compose.ui.extensions.toggleAlpha
 import io.getstream.video.android.core.call.state.CallAction
 
 /**
